Extract mockSql to be reusable by other tests
