+++ src/api.v2/model/experiment.js
@@ -54,6 +54,9 @@ const updateSamplePosition = async (experimentId, oldPosition, newPosition) => {
   const trx = await sql.transaction();
 
   try {
+    console.log('trx(tableName)');
+    console.log(await trx(tableName));
+
     const result = await trx(tableName).update({
       samples_order: trx.raw(`(
         SELECT jsonb_insert(samples_order - ${oldPosition}, '{${newPosition}}', samples_order -> ${oldPosition}, false)
