+++ tests/api.v2/model/experiment.test.js
@@ -1,31 +1,11 @@
 // Disabled ts because it doesn't recognize jest mocks
 // @ts-nocheck
-const _ = require('lodash');
-
 const sqlClient = require('../../../src/sql/sqlClient');
 const helpers = require('../../../src/sql/helpers');
 
 const validSamplesOrderResult = ['sampleId1', 'sampleId2', 'sampleId3', 'sampleId4'];
-const mockRawResult = Promise.resolve({ isRawResult: true });
-const queryBuilder = {
-  where: jest.fn().mockReturnThis(),
-  select: jest.fn().mockReturnThis(),
-  update: jest.fn().mockReturnThis(),
-  orderBy: jest.fn().mockReturnThis(),
-  raw: jest.fn(() => mockRawResult),
-  returning: jest.fn(),
-};
 
-const mockTrx = jest.fn(() => queryBuilder);
-mockTrx.commit = jest.fn().mockReturnThis();
-mockTrx.rollback = jest.fn().mockReturnThis();
-
-_.merge(mockTrx, queryBuilder);
-
-const mockSqlClient = {
-  isMockSqlClient: true,
-  transaction: jest.fn(() => Promise.resolve(mockTrx)),
-};
+const { mockSqlClient, mockTrx, mockTrxQueryBuilder } = require('../mocks/getMockSqlClient')();
 
 jest.mock('../../../src/api.v2/helpers/generateBasicModelFunctions',
   () => jest.fn(() => ({ hasFakeBasicModelFunctions: true })));
@@ -70,7 +50,7 @@ describe('model/experiment', () => {
       ['params_hash', 'state_machine_arn', 'execution_arn'],
       'pipeline_type',
       'pipelines',
-      expect.objectContaining({ isMockSqlClient: true }),
+      mockSqlClient,
     );
 
     const firstParam = helpers.aggregateIntoJson.mock.calls[0][0];
@@ -88,12 +68,13 @@ describe('model/experiment', () => {
     mockTrx.returning.mockImplementationOnce(
       () => Promise.resolve([{ samples_order: validSamplesOrderResult }]),
     );
+    mockTrx.raw.mockImplementationOnce(() => 'resultOfSql.raw');
 
     await experiment.updateSamplePosition(mockExperimentId, 0, 1);
 
     expect(mockTrx).toHaveBeenCalledWith('experiment');
 
-    expect(mockTrx.update).toHaveBeenCalledWith({ samples_order: mockRawResult });
+    expect(mockTrx.update).toHaveBeenCalledWith({ samples_order: 'resultOfSql.raw' });
     expect(mockTrx.raw.mock.calls[0]).toMatchSnapshot();
     expect(mockTrx.where).toHaveBeenCalledWith('id', 'mockExperimentId');
     expect(mockTrx.returning).toHaveBeenCalledWith(['samples_order']);
@@ -104,12 +85,13 @@ describe('model/experiment', () => {
 
   it('updateSamplePosition rolls back if the result is invalid', async () => {
     mockTrx.returning.mockImplementationOnce(() => Promise.resolve([{ samples_order: null }]));
+    mockTrx.raw.mockImplementationOnce(() => 'resultOfSql.raw');
 
     await expect(experiment.updateSamplePosition(mockExperimentId, 0, 1)).rejects.toThrow('Invalid update parameters');
 
     expect(mockTrx).toHaveBeenCalledWith('experiment');
 
-    expect(mockTrx.update).toHaveBeenCalledWith({ samples_order: mockRawResult });
+    expect(mockTrx.update).toHaveBeenCalledWith({ samples_order: 'resultOfSql.raw' });
     expect(mockTrx.raw.mock.calls[0]).toMatchSnapshot();
     expect(mockTrx.where).toHaveBeenCalledWith('id', 'mockExperimentId');
     expect(mockTrx.returning).toHaveBeenCalledWith(['samples_order']);
@@ -122,12 +104,13 @@ describe('model/experiment', () => {
     mockTrx.returning.mockImplementationOnce(
       () => Promise.resolve([{ samples_order: validSamplesOrderResult }]),
     );
+    mockTrx.raw.mockImplementationOnce(() => 'resultOfSql.raw');
 
     await expect(experiment.updateSamplePosition(mockExperimentId, 0, 10000)).rejects.toThrow('Invalid update parameters');
 
     expect(mockTrx).toHaveBeenCalledWith('experiment');
 
-    expect(mockTrx.update).toHaveBeenCalledWith({ samples_order: mockRawResult });
+    expect(mockTrx.update).toHaveBeenCalledWith({ samples_order: 'resultOfSql.raw' });
     expect(mockTrx.raw.mock.calls[0]).toMatchSnapshot();
     expect(mockTrx.where).toHaveBeenCalledWith('id', 'mockExperimentId');
     expect(mockTrx.returning).toHaveBeenCalledWith(['samples_order']);
